requirements.txt
Flask==0.11.1
google-api-python-client
httplib2
